node-data: move attestation structs in their own module
