+++ node-data/src/ledger/header.rs
@@ -18,7 +18,7 @@ pub struct Header {
     pub seed: Seed,
     pub state_hash: Hash,
     pub event_hash: Hash,
-    pub generator_bls_pubkey: bls::PublicKeyBytes,
+    pub generator_bls_pubkey: PublicKeyBytes,
     pub txroot: Hash,
     pub gas_limit: u64,
     pub iteration: u8,
@@ -103,7 +103,7 @@ impl Header {
             gas_limit,
             prev_block_hash,
             seed: Seed::from(seed),
-            generator_bls_pubkey: bls::PublicKeyBytes(generator_bls_pubkey),
+            generator_bls_pubkey: PublicKeyBytes(generator_bls_pubkey),
             iteration,
             state_hash,
             event_hash,
